[Sticky] Stickies now stick and translate w/ target
Some things: Z-index needs to be higher than the player for the sprite to display correctly, z-index standard should be added for stacking projectiles and other stuffs that needs to be drawn over player. Also sprites move weirdly and jerkily when attached player moves around (somethin up with lerp?). Getting started on EntityExplosion.